XLuaHook::invoke() must copy all Varargs elements used by the hook.
The Varargs object passed to invoke() seems to be reused by the Lua
interpreter and capturing it as is when defining the hook leads to
accessing other elements than intended when the hook runs.
Instead, copy all the function object in fun and the the extra
arguments in xargs, and use these when defining the hook.